CM-20 refactor: 여행지 폴더 구조 수정
