+++ src/main/java/net/coursemaker/backendv2/destination/command/domain/exception/DestinationNotFoundException.java
@@ -1,4 +1,4 @@
-package net.coursemaker.backendv2.destination.command.application.exception;
+package net.coursemaker.backendv2.destination.command.domain.exception;
 
 public class DestinationNotFoundException extends RuntimeException {
 	public DestinationNotFoundException(String message) {
